Changed file entry type to a pair of int64s rather than int32s. B-tree and file table in Fs_data now store pointers as int64s. Removed n_pointers and last_block_index functions from file.ml.

+++ lib/entry.ml
@@ -1,6 +1,6 @@
 type t = Tag.t * Cstruct.t
 type link = | Metadata of (int64 * int64)
-            | Data of (int32 * int32)
+            | Data of (int64 * int64)
 
 let sizeof t =
   Cstruct.length (snd t) + Tag.size
